moved the determination of sunrise and sunset times to a seperate script
