+++ crontab_scripts/moths_schedule.py
@@ -1,34 +1,10 @@
 #!/usr/bin/env python3
 #-*- coding: utf-8 -*-
 
-import subprocess
 import json
 from pathlib import Path
-import os
-from datetime import datetime, timedelta
+from datetime import datetime
 from crontab import CronTab
-
-def get_sunset_sunrise_times(latitude, longitude, date):
-    # Construct the Helicron command
-    # heliocron -date 2020-02-25 --latitude 52.752845 --longitude -3.253449 report --json
-    command = ["/home/pi/scripts/heliocron", "--date", date.strftime("%Y-%m-%d"), "--latitude", str(latitude), "--longitude", str(longitude), "report", "--json"]
-
-    # Run the command
-    result = subprocess.run(command, capture_output=True, text=True)
-    # print(result)
-
-    # Check if the command was successful
-    if result.returncode == 0:
-        # Split the output by newlines and return civil dawn and civil dusk times
-        stdout = json.loads(result.stdout)
-        #print(stdout)
-        sunset = datetime.strptime(stdout['sunset'], '%Y-%m-%dT%H:%M:%S%z')
-        sunrise = datetime.strptime(stdout['sunrise'], '%Y-%m-%dT%H:%M:%S%z')
-        return sunset.replace(tzinfo=None), sunrise.replace(tzinfo=None)
-    else:
-        # Handle error
-        print("Error:", result.stderr)
-        return None, None
 
 # Read the config file
 config_path = Path('/home/pi/config.json')
@@ -38,13 +14,13 @@ with config_path.open() as fp:
 # Get the camera setting from the config file
 camera_settings = config["camera_operation"]
 
-# Find today's and tomorrow's dates
-today = datetime.now()
-tomorrow = today + timedelta(1)
-
 # Get sunset and sunrise for today and tomorrow
-sunset, _ = get_sunset_sunrise_times(config['device_settings']['lat'], config['device_settings']['lon'], today)
-_, sunrise = get_sunset_sunrise_times(config['device_settings']['lat'], config['device_settings']['lon'], tomorrow)
+sunset_str = config["sunrise_sunset_times"]["sunset_time"]
+sunrise_str = config["sunrise_sunset_times"]["sunrise_time"]
+
+# Convert to datetime object
+sunset = datetime.strptime(sunset_str, '%H:%M:%S')
+sunrise = datetime.strptime(sunrise_str, '%H:%M:%S')
 
 # Access the user crontab
 ami_cron = CronTab(user='pi')
